MDL-77625: add a unit test to cover the bug
The test is intended to pass when the bug is fixed.
For now it fails because it asserts that questions are copied instead
of linked even after a repeated course backup/import operation.

The original author of this test is [Julien](https://tracker.moodle.org/browse/MDL-77625?focusedId=1074383&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-1074383). I just adjusted and refactored the test a bit before committing to git.